feat: created an independent secret referencing module and update imports
